Make xclip exit when it loses the selection
1.Fix the issue #64(https://github.com/astrand/xclip/issues/64).

When receiving a SelectionClear event (i.e; lose the selection), the while loop for checking
dloop as-is ran endlessly if sloop is set to -1 (a.k.a unlimited number of requests).
This caused xclip fail to exit when it loses the selection.

Fix the problem via checking whether there is any in-progress transfer. If no, force xclip exit.

2.Add a test case in xctest to make sure xclip can exit properly. (credit: hackerb9)