Set sloop instead of dloop to exit when losing selection
Previously, we set dloop=sloop when we lost the selection so that,
once all the requestors were finished, xclip wouldn't loop around
again. The problem is that the loop conditional was not simply
`if (dloop < sloop)`, but  `if (dloop<sloop || sloop<1)`.

The reason for that is that sloop == 0 means keep looping infinitely.

The fix: change the conditional to `if (dloop<sloop || sloop==0)`
which is not only better style as it explicitly checks for the flag to
loop infinitely, but allows us to set sloop to a negative number to
prevent further looping.

This can be tested with the 'borked 1' test.